добавил метод для обнуления карты перед тестами

+++ src/main/java/com/system/repositories/HltDispCardRepository.java
@@ -4,7 +4,9 @@ import com.system.model.HltDispCardEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.UUID;
+
 @Repository
 public interface HltDispCardRepository extends JpaRepository<HltDispCardEntity, Long> {
-    HltDispCardEntity getByDispCardId(int number);
+    HltDispCardEntity getByDispCardId(long number);
 }
